REFACT: move users.type.ts 's Meta interface file
